REST 21:04/requirements.txt
asgiref==3.6.0
asn1crypto==1.5.1
backports.zoneinfo==0.2.1
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
defusedxml==0.7.1
Django==4.1.7
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
et-xmlfile==1.1.0
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.1
oauthlib==3.2.2
openpyxl==3.0.10
pandas==1.5.3
pg8000==1.29.4
Pillow==9.5.0
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
scramp==1.4.4
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.1
SQLAlchemy==2.0.7
sqlparse==0.4.3
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.15.0
